+++ src/pages/ResultPage.css
